requirements.txt
pyyaml
python-gitlab
